test(shared): relocate and scaffold test structure for schemas and db
