+++ shared/__tests__/db/db.test.ts
@@ -1,0 +1,6 @@
+import { assert as _assert } from "@std/assert";
+// import { dbClient } from "../db/db.client.ts";
+
+Deno.test("Shared DB", () => {
+  // TODO: Test shared DB logic
+});
