+++ shared/__tests__/schemas/api/common.test.ts
@@ -1,5 +1,5 @@
 import { assert as _assert } from "@std/assert";
-import { QuerySchema as _QuerySchema } from "../../../../../shared/schemas/api/common.ts";
+import { QuerySchema as _QuerySchema } from "../../../schemas/api/common.ts";
 
 Deno.test("QuerySchema", () => {
   // TODO: Test valid input parsing
