+++ shared/__tests__/schemas/api/payments.test.ts
@@ -1,5 +1,5 @@
 import { assert as _assert } from "@std/assert";
-import { CreatePaymentSchema as _CreatePaymentSchema } from "../../../../../shared/schemas/api/payments.ts";
+import { CreatePaymentSchema as _CreatePaymentSchema } from "../../../schemas/api/payments.ts";
 
 Deno.test("CreatePaymentSchema", () => {
   // TODO: Test valid input parsing
