+++ shared/__tests__/schemas/api/people.test.ts
@@ -1,5 +1,5 @@
 import { assert as _assert } from "@std/assert";
-import { CreateMemberSchema as _CreateMemberSchema, UpdateMemberSchema as _UpdateMemberSchema } from "../../../../../shared/schemas/api/people.ts";
+import { CreateMemberSchema as _CreateMemberSchema, UpdateMemberSchema as _UpdateMemberSchema } from "../../../schemas/api/people.ts";
 
 Deno.test("CreateMemberSchema", () => {
   // TODO: Test valid input parsing
