+++ shared/__tests__/schemas/api/pots.test.ts
@@ -1,5 +1,5 @@
 import { assert as _assert } from "@std/assert";
-import { CreatePotSchema as _CreatePotSchema, UpdatePotSchema as _UpdatePotSchema } from "../../../../../shared/schemas/api/pots.ts";
+import { CreatePotSchema as _CreatePotSchema, UpdatePotSchema as _UpdatePotSchema } from "../../../schemas/api/pots.ts";
 
 Deno.test("CreatePotSchema", () => {
   // TODO: Test valid input parsing
