+++ shared/__tests__/schemas/api/reservations.test.ts
@@ -1,5 +1,5 @@
 import { assert as _assert } from "@std/assert";
-import { CreateReservationSchema as _CreateReservationSchema } from "../../../../../shared/schemas/api/reservations.ts";
+import { CreateReservationSchema as _CreateReservationSchema } from "../../../schemas/api/reservations.ts";
 
 Deno.test("CreateReservationSchema", () => {
   // TODO: Test valid input parsing
