+++ shared/__tests__/schemas/api/tags.test.ts
@@ -1,5 +1,5 @@
 import { assert as _assert } from "@std/assert";
-import { CreateTagSchema as _CreateTagSchema, UpdateTagSchema as _UpdateTagSchema } from "../../../../../shared/schemas/api/tags.ts";
+import { CreateTagSchema as _CreateTagSchema, UpdateTagSchema as _UpdateTagSchema } from "../../../schemas/api/tags.ts";
 
 Deno.test("CreateTagSchema", () => {
   // TODO: Test valid input parsing
