+++ shared/__tests__/schemas/api/transactions.test.ts
@@ -1,5 +1,5 @@
 import { assert as _assert } from "@std/assert";
-import { CreateTransactionSchema as _CreateTransactionSchema, UpdateTransactionSchema as _UpdateTransactionSchema } from "../../../../../shared/schemas/api/transactions.ts";
+import { CreateTransactionSchema as _CreateTransactionSchema, UpdateTransactionSchema as _UpdateTransactionSchema } from "../../../schemas/api/transactions.ts";
 
 Deno.test("CreateTransactionSchema", () => {
   // TODO: Test valid input parsing
